Pacify Coverity (CID #1690435)

+++ src/lib/io/coord.c
@@ -759,10 +759,10 @@ int fr_coord_to_worker_send(fr_coord_t *coord, int32_t worker_id, uint32_t cb_id
  */
 int fr_coord_to_worker_broadcast(fr_coord_t *coord, uint32_t cb_id, fr_dbuff_t *dbuff)
 {
-	uint32_t	i;
+	int32_t		i;
 	int		failed = 0;
 
-	for (i = 0; i < coord->max_workers; i++) {
+	for (i = 0; i < (int32_t)coord->max_workers; i++) {
 		if (!coord->coord_send_control[i - MIN_WORKER_ID]) continue;
 		if (fr_coord_to_worker_send(coord, i, cb_id, dbuff) < 0) failed++;
 	}
